started separating terminus-electron and terminus-web

+++ scripts/prepackage-plugins.js
@@ -11,6 +11,9 @@ sh.mkdir('-p', target)
 fs.writeFileSync(path.join(target, 'package.json'), '{}')
 sh.cd(target)
 vars.builtinPlugins.forEach(plugin => {
+  if (plugin === 'terminus-web') {
+    continue
+  }
   log.info('install', plugin)
   sh.cp('-r', path.join('..', plugin), '.')
   sh.rm('-rf', path.join(plugin, 'node_modules'))
